requirements.txt
streamlit>=1.30.0
boto3>=1.34.0
streamlit-authenticator>=0.2.3
python-dotenv>=1.0.0
jinja2>=3.1.2
pandas>=2.0.0
pyyaml>=6.0
msal>=1.25.0
